pyproject.toml
[project]
name = "postprocessing-and-gaze"
version = "0.1.0"
description = "Add your description here"
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "scipy>=1.15.2",
    "tensorboardx>=2.6.2.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
]
